[tacho-motor] Remove `type` attribute and replace with `driver_name`.
Also includes adding LEGO_DRIVER_NAME property and removing
speed_regulation_K attribute.

This is how we are implementing other drivers. There is a lookup table and a
_defs.c file that has all of the model specific driver info.

Also fixed issue with position attribute implemtation so that it can now return
an error if needed.

Issue ev3dev/ev3dev#282