[tacho-motor] Rename pulses_per_second to speed.
Modified the ops so that getting and setting speed and speed_sp can return an
an error. Also fixed up duty_cycle and duty_cycle_sp the same for returning
an error.

There was a duplicate speed variable in legoev3_motor that was never set, but
used in an if statement, so may introduce a bug or fix a bug.

Issue ev3dev/ev3dev#282