[tacho-motor] Move run-timed command handling to tacho-motor class.
Previously, it was handled by the implementing driver, but there is no need for
this. If a motor can run forever and stop, then it can run timed.

Also fixed a few TODOs and added checks to prevent getting in an invalid state
in the NxtMMX driver.

Issue ev3dev/ev3dev#282