[dc-motor] Update to be more like tacho-motor class.
This makes the dc-motor class look and feel more like the tacho-motor class.
Parameters now only take effect when a command is sent. The command attribute
is now write-only. There is a state attribute that returns flags instead of
reading the command attribute. There are separate stop_command and stop_commands
attributes to control how the motor stops. Also added a run-timed command.

Internally, there is now more of a delay between ramping updates. This will allow
the dc-motor class to be used with controllers that have slow communication.
Also simplified things a bit by removing the direction field.