[lego-sensor] Add `direct` attribute.
The direct attribute is uses for directly reading or writing data to/from
a sensor that supports it. For example with I2C sensors, this allows you to
read or write to an arbitrary register.

Writing was previously possible with the `bin_data` attribute, but this
has been removed and the behavor has been slightly changed.