[sensors] Remove 50 ms hard limit on poll_ms
Also use hrtimer for polling for better resolution. This will allow sensors to
be polled as fast as possible. Unfortunatly, this can cause the autodetection
to fail on EV3 input ports, which causes the sensor to disconnect and reconnect.
This probably cannot be fixed in software because there needs to be some "quiet
time" in order for autodetection to work.