PiStorms: Pass in_port to nxt_i2c_sensor driver.
This allows the driver to display the expected port name. This is kind of
cheating since the i2c is just passed through the ports, so with the wrong
conditions, it is possible to get the wrong port name on an i2c device. But
for most cases (i.e. one sensor per port), this should work fine.