Rename port_name to address.
This is a major breaking change. It affects sysfs attributes and udev
properties for all device types.

Originally, there were separate port_name and address attributes. A while
back, these were merged into the port_name attribute. This makes the name
"port_name" a bit misleading. The name may be a port name or have multiple
port names or contain other information, like I2C addresses. The combination
of all of this information can still be called an "address" since it tells
you where a device is. Also, using port_name can be confusing because there
may or may not be a lego-port class device with the same port_name.

This commit also combines the two LEGO_NAME_SIZE and LEGO_PORT_NAME_SIZE
constants that were the same value.