evb: Add sound driver.
This is mostly copied from legoev3_sound.c. However, the BeagleBone does
not allow you to use FIQ, so the FIQ parts have been removed. Furthermore,
the standard kernel pwm drivers do not expose PWM interrupts, so we are
resorting to using hrtimers for interrupts. The sound is quite good when
the system is not busy, but starts getting scratchy when the system starts
getting busy.

It also features separate volume controls for PCM and Beep and device-tree
support.

Since we can't use FIQ, it would be nice to eventually use the PRU for PCM
playback. However, since the PRU doesn't have muliply and divide instructions,
we will have to figure out how to prescale the PCM data in the CPU.