evb: Add support for input and output ports.
The ports are nearly identical to the EV3. Code is copied and updated to
use device tree instead of platform_data.

One significat difference is that the EVB does not have a gpio on input port
pin 2, so NXT sensors cannot be properly detected. There is also no 9V out
on input port pin 1, but this is not a big deal since it is only used by the
NXT Ultrasonic sensor.  Also, support for I2C on the input ports is not
implemented yet.