brickpi: Fix kernel oops when setting uart sensor
When a new sensor device is created, info->sensor was not set in the
brickpi_in_port_set_ev3_uart_sensor_mode. The first call worked because
the ev3_uart_sensor_probe was deferred because the module was loading,
but failed the second time because the set_mode is called in the probe
function.

Worked around by passing the necessary information to the callback.

Issue ev3dev/ev3dev#531