motors: use driver info instead of callback functions
This changes the tacho-motor-class to use info from the motor defs instead
of relying on callbacks for static motor information. The info stuct is now
passed to the tacho-motor-class when a device is registered.

Had to divide position by 2 on BrickPi since BrickPi counts rising and
falling edgeas one count each instead of one count total.