motors: Add tacho motor helpers
This adds a new file to the tacho_motor module that has helper functions
that can optionally be used by tacho motor drivers. The idea is that if
the hardware does not implement a function and something has to be
implemented in software, then it should be here so that it can be shared
by all tacho motor drivers.

The first set of helper functions is motor speed calculation based on
position. For motor controllers like BrickPi and PiStorms, the speed is
not availible and must be calculated. On both of these, we only have the
position to go by and it is polled at a fixed interval, finding the change
in position over change in time is the best we can do. It will be a bit
inaccurate a low speeds, but should be "good enough".