legoev3_motor: don't track tacho motor command
It doesn't make sense to track the higher level tacho motor command anymore.
It just makes it confusing because the active command in legoev3_motor will
not necessaraly match the active command in tacho_motor_class.

The only thing this was used for anymore was to tell if we are running to a
position, so replacing it with a bool to track that.