tacho-motor: re-implement polarity
Polarity was removed from the lower level motor drivers, so re-implementing
it in the tacho motor class. Changing polarirty will take effect immediatly
but will only affect the sign of the values stored and returned by sysfs
attributes. It will not change the stored values themselves, which would
cause problems with motor control if it did.

For example if the position is 123 and the speed is 456, then inverting polarity
will cause the position to be read as -123 and the speed as -456.

Issue ev3dev/ev3dev#601
Issue ev3dev/ev3dev#545