pixycam: drop COL-CODE and ANGLE modes
Reading color code info requires sending the color code to the pixycam
which cannot be done by setting a mode (it is a 16-bit value). Also,
the ANGLE mode doesn't really make sense. The value is only updated
when the ALL mode is active, so you constantly have to swith modes to
get a value.

Issue ev3dev/ev3dev#652