ev3: add other-i2c mode to input ports
This adds a new "other-i2c" mode to input ports on the EV3. This mode
configures the ports for I2C communications, but unlike the "nxt-i2c"
mode, it does not set an adapter class, so it does not probe the port
for any sensors.

Issue ev3dev/ev3dev#649