lms2012: slow down ADC polling by factor of 10
We are currently using a bit-banged SPI for the ADC chip. This is very
CPU intensive. The driver was trying to poll all analog channels at 1kHz
which made the whole system very slow. By slowing this down by a factor
of 10, we free up the CPU for other stuff. Polling at 1kHz only matters if
we are trying to log an analog sensor that fast. For everything else, 100Hz
will be plenty fast.