lms2012: d_uart: update to use line discipline
The d_uart driver was poking UART registers directly, which meant it only
worked for BeagleBone hardware. This removes the hardware-specific code
and replaces it with a line discipiline so that we can use any serial port
without special handling.

There are a few issues here where we are calling sleeping functions in an
atomic context, but it seems to be working, so we will leave it for the
time being.

Also was experiencing some lock ups before changing the hrtimer to use a
tasklet. Hopefully that fixed the issue.

One down side of this is that userspace will have to have a udev rule to
run ldattach on all of the serial ports. But, there is not any option for
in kernel users of ttys other than this, so we will have to live with it.