lms2012: d_analog: Increase delay for detecting output port pin 5 change
Since we are polling the AD/C slower than the official firmware, this value
does not get updated quite as fast. So, the device connection manager needs
to wait longer before reading the output pin 5 analog value.

This fixes the medium motor not being correctly detected when the shaft is
in a position that causes pin 5 to be pulled high.

Issue: ev3dev/lms2012-compat#47