lms2012: d_pwm: Don't use platform-specific MAX_PWM_CNT
The value MAX_PWM_CNT is based on the 132MHz PWM clock on the EV3. Since
this driver uses the generic pwm framework now, this value is no longer
meaningful, so we are using PWM_PERIOD instead.

Since this value is larger, there were some overflow issues when doing the
scaling for adjusting for minimum value, so a new inline function was added
to take care of this more generically.