evb: ports: invert pin 6 in output port connection manager
In device tree, we are using GPIO_ACTIVE_LOW for pins 5 and 6 because
there are inverting buffers connected to both pins. This way, the gpio
will reflect the actual external pin state. So, we have to take this
into account in the device connection manager. The official firmware
source that this is based on does not invert the gpios.

Issue: ev3dev/ev3dev#809