brickpi3: add support for stacking multiple boards
BrickPi3 supports stacking multiple boards, so we accomidate this by
adding an address parameter everywhere for the board address. The
previous assumption that the address was the chip select was incorrect.

Most things have been tested working, however the battery driver is not
aware of multiple boards. We will either have to do some more device
tree magic or register the platform device in code to accomodate this.