legoev3_motor: add separate flag for position_sp
This adds a new flag to indicate what should be used for the position
PID. If we are running to a position and it completes without interruption,
we should use the position_sp set in sysfs. Otherwise, we should use the
current position, e.g. if stop action is sent before we reach the final
position.

Issue: ev3dev/ev3dev#847