ev3: motor: replace spinlock with rwsem
since there is nothing left running in an interrupt context, we no
longer need to disable interrupts in critical sections. usinng a read/
write semaphore is good enough. also protect some read sections while we
are at it, which is especially important when we are reading more than
one value.