brickpi3: add delay when scheduling first work callback in run-to-pos
If we don't have a delay, when we call brickpi3_run_to_pos_work(), the
speed can still be 0. If the current position is < 10 counts from the
position setpoint, this will cause the motor stop before it has reached
the setpoint. 20ms should be plenty of time for the speed to change.

Issue: ev3dev/ev3dev#1237