sensors/ev3_uart_sensor_ld: improve resync of data messages
It is common to get UART hardware buffer overruns and miss data after
switching to the higher baud rate for receiving data. This adds some
extra checks to try to get back into sync after this happens without
having to wait for the sensor to timeout and go through the connection
sequence again. It doesn't completely fix the problem but is an
improvement over the current situation.